backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.0",
    "google-cloud-speech>=2.38.0",
    "httpx>=0.28.1",
    "instaloader>=4.15.1",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.45.0",
    "yt-dlp>=2026.3.17",
]
